backend/requirements.txt
# MindSupport Backend Dependencies
# Using versions with pre-built wheels to avoid Rust compilation issues

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
motor>=3.3.0
pymongo>=4.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation & Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# OpenAI
openai>=1.0.0

#gemini
google-genai>=1.0.0

# CORS
python-dotenv>=1.0.0

# WebSocket (included in fastapi)
websockets>=12.0
